# Crop incoming lines to Area of Interest