… scripts/benchmarking_input_data/RasterLinesJoin_Schuylkill_HUC08.json