@kellyi 2017-11-15
Q: instead of using the scripts above will the existing `./scripts/benchmark` work just as well for testing purposes?