@kellyi 2017-11-15
Nevermind, looks like we don't have a benchmarking test on RasterLinesJoin? Wonder if it's worth adding one.