@rajadain 2017-11-15
I added the latest `RasterLinesJoin` to the benchmark script with the old behavior, then added my fix, then updated the benchmark script with the new behavior, if you'd like to use that instead.