@rajadain 2017-11-15
The order of commits in GitHub is incorrect (because it uses Author date, not Commit date). The true order is this:

```
* 118eeea - (HEAD -> tt/crop-lines-to-aoi, origin/tt/crop-lines-to-aoi) Update RasterLinesJoin benchmark (66 seconds ago) <Terence Tuhinanshu>
* e9e0aa6 - Crop incoming lines to area of interest (83 seconds ago) <Terence Tuhinanshu>
* 1818cb8 - Add RasterLinesJoin benchmark (83 seconds ago) <Terence Tuhinanshu>
*   de66670 - (origin/develop, origin/HEAD, develop) Merge pull request #75 from WikiWatershed/ki/improve-error-reporting (8 weeks ago) <Kelly Innes>
```