@kellyi 2017-11-15
RLJ timings on `develop`:

```sh
Timing HUC8 RasterLinesJoin ->

HUC8 RasterLinesJoin, run 1 -> 6.944408 s
HUC8 RasterLinesJoin, run 2 -> 1.918769 s
HUC8 RasterLinesJoin, run 3 -> 2.046422 s
HUC8 RasterLinesJoin, run 4 -> 1.342596 s
HUC8 RasterLinesJoin, run 5 -> 1.817221 s

HUC8 RasterLinesJoin average -> 2.8138832 s
```