@kellyi 2017-11-15
This is probably a more accurate RLJ run from `develop` since it leaves off the initial 7 second warmup run:

```
Timing HUC8 RasterLinesJoin ->

HUC8 RasterLinesJoin, run 1 -> 2.000698 s
HUC8 RasterLinesJoin, run 2 -> 1.803359 s
HUC8 RasterLinesJoin, run 3 -> 1.714622 s
HUC8 RasterLinesJoin, run 4 -> 1.753615 s
HUC8 RasterLinesJoin, run 5 -> 1.450551 s

HUC8 RasterLinesJoin average -> 1.744569 s
```