@kellyi 2017-11-15
Here's the benchmark timing from this branch after a few runs:

```
Timing HUC8 RasterLinesJoin ->

HUC8 RasterLinesJoin, run 1 -> 1.379907 s
HUC8 RasterLinesJoin, run 2 -> 1.380688 s
HUC8 RasterLinesJoin, run 3 -> 1.442743 s
HUC8 RasterLinesJoin, run 4 -> 1.693359 s
HUC8 RasterLinesJoin, run 5 -> 1.430437 s

HUC8 RasterLinesJoin average -> 1.4654268 s
```