@kellyi 2017-11-15
> Ensure the two output files are different.

Output differs:

```
$ diff develop-output.json develop-clipped-streams.json
3,5c3,5
<         "List(42)": 2149,
<         "List(22)": 3252,
<         "List(43)": 2387,
---
>         "List(42)": 2215,
>         "List(22)": 3253,
>         "List(43)": 2393,
7,8c7,8
<         "List(41)": 33341,
<         "List(21)": 11553,
---
>         "List(41)": 33404,
>         "List(21)": 11558,
12,14c12,14
<         "List(90)": 14005,
<         "List(52)": 2874,
<         "List(11)": 8095,
---
>         "List(90)": 14014,
>         "List(52)": 2876,
>         "List(11)": 8128,
16,17c16,17
<         "List(82)": 4736,
<         "List(81)": 10301,
---
>         "List(82)": 4745,
>         "List(81)": 10314,
```