@kellyi 2017-11-15
New output from this branch matches the `develop-output.json` above:

![screen shot 2017-11-15 at 5 50 36 pm](https://user-images.githubusercontent.com/4165523/32864540-93a98c0a-ca2d-11e7-898c-a9a8ec02a341.png)