@kellyi 2017-11-15
This seems like it's good to go in: it works as described by the testing instructions, and the benchmarking scripts don't show much much difference in the timings. Makes me wonder whether there are other operations we're doing Python-side that could be sped up by moving them here.