@mmcfarland 2017-11-16
I apologize that this comment comes from a place of limited context, but how does this relate to the existing work done in https://github.com/WikiWatershed/model-my-watershed/pull/2506?  My concern about the comparison between PostGIS vs this branch is that it may have been done in the context of VM resources in the development environment.  This may indeed be an improvement, but I'd like to see a more detailed description of the decision points and observations that led to this changeset, and how it relates to the previous work proposed.