@rajadain 2017-11-16
Spoke with @mmcfarland in person about the motivations and effects, and we have a green light. I'm going to merge this and make a new release, then include it in WikiWatershed/model-my-watershed#2502.